release projects/Signing_Server/requirements.txt
fastapi
uvicorn
cryptography
pycryptodome
